+++ rustls/src/error.rs
@@ -445,28 +445,6 @@ impl PartialEq<Self> for CertRevocationListError {
     }
 }
 
-impl From<webpki::Error> for CertRevocationListError {
-    fn from(e: webpki::Error) -> Self {
-        use webpki::Error::*;
-        match e {
-            InvalidCrlSignatureForPublicKey
-            | UnsupportedCrlSignatureAlgorithm
-            | UnsupportedCrlSignatureAlgorithmForPublicKey => Self::BadSignature,
-            InvalidCrlNumber => Self::InvalidCrlNumber,
-            InvalidSerialNumber => Self::InvalidRevokedCertSerialNumber,
-            IssuerNotCrlSigner => Self::IssuerInvalidForCrl,
-            MalformedExtensions | BadDer | BadDerTime => Self::ParseError,
-            UnsupportedCriticalExtension => Self::UnsupportedCriticalExtension,
-            UnsupportedCrlVersion => Self::UnsupportedCrlVersion,
-            UnsupportedDeltaCrl => Self::UnsupportedDeltaCrl,
-            UnsupportedIndirectCrl => Self::UnsupportedIndirectCrl,
-            UnsupportedRevocationReason => Self::UnsupportedRevocationReason,
-
-            _ => Self::Other(Arc::new(e)),
-        }
-    }
-}
-
 impl From<CertRevocationListError> for Error {
     #[inline]
     fn from(e: CertRevocationListError) -> Self {
@@ -592,58 +570,6 @@ mod tests {
         let other = Other(alloc::sync::Arc::from(Box::from("")));
         assert_ne!(other, other);
         assert_ne!(BadSignature, InvalidCrlNumber);
-    }
-
-    #[test]
-    fn crl_error_from_webpki() {
-        use super::CertRevocationListError::*;
-        let testcases = &[
-            (webpki::Error::InvalidCrlSignatureForPublicKey, BadSignature),
-            (
-                webpki::Error::UnsupportedCrlSignatureAlgorithm,
-                BadSignature,
-            ),
-            (
-                webpki::Error::UnsupportedCrlSignatureAlgorithmForPublicKey,
-                BadSignature,
-            ),
-            (webpki::Error::InvalidCrlNumber, InvalidCrlNumber),
-            (
-                webpki::Error::InvalidSerialNumber,
-                InvalidRevokedCertSerialNumber,
-            ),
-            (webpki::Error::IssuerNotCrlSigner, IssuerInvalidForCrl),
-            (webpki::Error::MalformedExtensions, ParseError),
-            (webpki::Error::BadDer, ParseError),
-            (webpki::Error::BadDerTime, ParseError),
-            (
-                webpki::Error::UnsupportedCriticalExtension,
-                UnsupportedCriticalExtension,
-            ),
-            (webpki::Error::UnsupportedCrlVersion, UnsupportedCrlVersion),
-            (webpki::Error::UnsupportedDeltaCrl, UnsupportedDeltaCrl),
-            (
-                webpki::Error::UnsupportedIndirectCrl,
-                UnsupportedIndirectCrl,
-            ),
-            (
-                webpki::Error::UnsupportedRevocationReason,
-                UnsupportedRevocationReason,
-            ),
-        ];
-        for t in testcases {
-            assert_eq!(
-                <webpki::Error as Into<CertRevocationListError>>::into(t.0),
-                t.1
-            );
-        }
-
-        assert!(matches!(
-            <webpki::Error as Into<CertRevocationListError>>::into(
-                webpki::Error::NameConstraintViolation
-            ),
-            Other(_)
-        ));
     }
 
     #[test]
