+++ rustls/src/key.rs
@@ -1,7 +1,5 @@
 use core::fmt;
 
-use crate::Error;
-
 /// This type contains a private key by value.
 ///
 /// The private key must be DER-encoded ASN.1 in either
@@ -96,32 +94,5 @@ impl fmt::Debug for Certificate {
         f.debug_tuple("Certificate")
             .field(&BsDebug(&self.0))
             .finish()
-    }
-}
-
-/// wrapper around internal representation of a parsed certificate. This is used in order to avoid parsing twice when specifying custom verification
-#[cfg_attr(not(feature = "dangerous_configuration"), allow(unreachable_pub))]
-#[cfg_attr(docsrs, doc(cfg(feature = "dangerous_configuration")))]
-pub struct ParsedCertificate<'a>(pub(crate) webpki::EndEntityCert<'a>);
-
-impl<'a> TryFrom<&'a Certificate> for ParsedCertificate<'a> {
-    type Error = Error;
-    fn try_from(value: &'a Certificate) -> Result<ParsedCertificate<'a>, Self::Error> {
-        webpki::EndEntityCert::try_from(value.0.as_ref())
-            .map_err(crate::verify::pki_error)
-            .map(ParsedCertificate)
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use super::Certificate;
-
-    #[test]
-    fn certificate_debug() {
-        assert_eq!(
-            "Certificate(b\"ab\")",
-            format!("{:?}", Certificate(b"ab".to_vec()))
-        );
     }
 }
